backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
requests
beautifulsoup4
